requirements.txt
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu
llama-cpp-python==0.2.86
huggingface-hub>=0.24.5
